fix: entity 수정

+++ loopz-product-service/src/main/java/kr/co/loopz/dto/response/ObjectResponse.java
@@ -6,7 +6,7 @@ public record ObjectResponse (
         String objectName,
         String intro,
         String imageUrl,
-        int objectPrice,
+        Long objectPrice,
         boolean soldOut,
         Boolean liked
 ){
